requirements.txt
# Core libraries
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
tqdm>=4.41.0

# NNI for hyperparameter tuning
nni>=2.0

# Visualization and analysis
matplotlib>=3.2.0
seaborn>=0.11.0
pandas>=1.1.4

# AWS and S3 support
boto3>=1.24.0
botocore>=1.27.0

# Optional: for deployment
# fastapi>=0.68.0
# uvicorn>=0.15.0
# python-multipart>=0.0.5